(fix) Remove attribute binding on `value`
Because of this attribute binding, assigning a date to the component's `value` attribute will cause Ember to assign the date to the underlying `<input>`'s value attribute, which is in the best case useless.

Flatpickr will then overwrite this value, except in some cases, where the visible string will become the Date instance cast to a string (i.e. default JavaScript date string formatting).